requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
requests==2.31.0
python-dotenv==1.0.1
unidiff==0.7
pydantic==1.10.9
httpx==0.24.1
mypy_extensions==0.4.3